+++ src/App.js
@@ -1,9 +1,9 @@
-import Input from "./components/Input";
+import Layout from "./components/Layout";
 
 function App() {
   return (
     <div className="App">
-      <Input />
+      <Layout />
     </div>
   );
 }
