+++ src/components/Movies.jsx
@@ -1,5 +1,4 @@
-import axios from "axios";
-import React, { useState, useContext } from "react";
+import React, { useContext } from "react";
 import StoreCtx from "../store/store-context";
 import BasicMovieList from "./BasicMovieList";
 
@@ -9,30 +8,10 @@ import "./styles/movies.css";
 const Movies = (props) => {
   const ctx = useContext(StoreCtx);
 
-  const [idOfMovie, setIdOfMovie] = useState(null);
-
-  let filtered = null;
-
-  if (idOfMovie) {
-    filtered = ctx.movies.filter((item) => item.id === idOfMovie);
-  }
-
   return (
     <>
-      {!ctx.showModal && (
-        <BasicMovieList
-          setIdOfMovie={setIdOfMovie}
-          fetchDetailsFromWiki={ctx.fetchDetailsFromWiki}
-        />
-      )}
-
-      {ctx.showModal && idOfMovie && (
-        <MovieDetails
-          details={ctx.detailsFromWiki}
-          searchKey={ctx.wikiSearchKey}
-          filtered={filtered}
-        />
-      )}
+      {!ctx.showModal && <BasicMovieList />}
+      {ctx.showModal && ctx.idOfMovie && <MovieDetails />}
     </>
   );
 };
